Restore pre-init hardware state values
The applications offers the option to not control some system
components. Ideally, the disabled system components should not be
touched by the application. However, in order to obtain valid default
values for some controls, the application must submit reset commands
to some hardware features on its startup.

This patch allow the restoration of the pre-init harware values (when
supported [1]) in order to maintain the previous state of disabled
components.

[1] At the moment, there is no way to restore the enabled states
through pp_dpm_* files. In this case, all states will be enabled when
the application starts.

Changes:
- Reduce the controls that submit reset commands to the ones
that initializes their internal values with the hardware
defaults (PMFVState, PMFVVoltCurve and PMPowerCap).
- Add a post-init phase that restore the pre-init hardware
values that were reseted by some controls on pre-init.