Only refresh graph series when the graph is visible
GraphItem points are updated as before but now no more cpu cycles are
wasted on replacing the points of the line series when the graph is
not visible.

Also, propagate visibility of the main window to its components,
so it works as expected when the window is closed on the system tray.