Fix missing character on profile menu icon
For some reason, Qt fails to find the used character (\uFE19) even
when the font that includes it is installed on the system (usually on
DroidSansFallback).

Use \u22EE instead, as it seems that this one is included on more
fonts.

(cherry picked from commit 928dc4b6926c92bad7da4d4131bb23d9a9fa820d)