Implement sensor ignoring on the system model
SysModelSyncer uses a list of ignored sensors, gathered from the
'value' parameter of SysModelSyncer::settingChanged method. This
method parses the list of sensors to be ignored when the changed
setting key is 'Workaround/ignoredSensors'.

The list of ignored sensors are passed to the system model by
SysModelSyncer everytime the system model updates its sensors. The
list consists of a map of sets of sensors ids indexed by the system
component id.

On the updateSensor method of every system component, the sensor is
skipped when is present on the list of ignored sensors.