Add deferred system tray icon creation
Used when the system tray is not available during SysTray
construction.

This can happen when the desktop has no support for system tray or
during desktop startup, when the application is launched before the
system tray initialization.

In the first case, the system tray icon won't be created. In the later
one, the system tray icon will be created if the system tray becomes
available within a specific period of time (60 seconds).

Fixes #43