Rework vega20+ advanced powermanagement control
On vega20 and later architectures, the states listed on
pp_od_clk_voltage are used to control the minimum and maximum allowed
frequencies on each component.

The control has been reworked to use the ranges and to not interfere
with the available states exposed through pp_dpm_sclk and
pp_dpm_mclk.

Fixes #37