Replace KAuth dependency with Polkit and Qt5::DBus
There are multiple problems with KAuth:

- KAuth generates D-Bus service files with a wrong path to the service
executable file when a custom installation prefix is used.

- KAuth also places D-Bus service config files into the installation
prefix, which can lead to a non working installation when using a custom
prefix and D-Bus has not been configured to check for the service config
files on the custom prefix.

This patch replaces KAuth with an implementation that uses Polkit and
the Qt D-Bus component.

Using Polkit and D-Bus over KAuth has the following advantages:

- D-Bus service configs are installed by default on the standard D-Bus
service datadir path, leading to working installations when custom
installation prefix is being used. This behaviour can be changed by
setting INSTALL_DBUS_FILES_IN_PREFIX to ON.

- D-Bus service files are now generated with the correct path to the
service executable file when a custom installation prefix is used.

- Helper service crashes are now handled. The service will be restarted
on the fly when needed.

- cmake-extra-modules dependency is removed as is no longer required.

Fixes #276.