Add support for disabling a frequency range bound on PMFreqRange
PMFreqRange control will be enabled even when a bogus frequency range
bound is detected. In this case, the control will only provide
functionality for the unaffected range bound.

This implementation provides a workaround for a long-standing driver bug
which affects the RX6000 series. It allows the user to also control the
maximum memory frequency on these cards.

See #213.