Support power profile mode on asics without heuristics settings
There are asics without heuristics settings. On these asics, the
description of the columns (located in the first line, as a header) is
also missing.

Both the functions and the regular expressions used to get the active
power profile and the list of power profile modes from
pp_power_profile_mode has been simplified. The header is no longer
checked and not matched lines are just ignored.

Fixes #303.