Use GPU unique IDs to identify GPUs within profiles
Using the index, the device ID and the revision number of the GPU to
uniquely identify GPUs within a profile doesn't work consistently. The
index depends on the GPU initialization order, but the kernel does not
guarantee that the same order is maintained between reboots. Therefore,
this identification method fails when different GPU models are used in
the same system.

Use GPU unique IDs when available to solve this issue. This new
identification method takes precedence over the older and faulty method,
which will be used when unique IDs are not stored on the profile or when
unique IDs are not supported by the hardware.

Fixes #300.