Revert "Add support for disabling a frequency range bound on PMFreqRange"
This reverts commit a821c5cab170d6a0379f12a2972c3ba722d03033.

Adjusting the maximum frequency on RX6000 series is triggering a driver
bug that locks the memory frequency to a low value.

As, this is probably related to the driver bug that produces #213, it's
safer to just disable the memory clocks controls altogether until AMD
developers fix the memory related RX6000 driver bugs once and for all.

Fixes #304.