Fix crash on application exit when running under some window managers
On some window managers (like IceWM), the system tray was crashing upon
destruction because it was parented to QApplication. As commented on
QTBUG-22632, QSystemTrayIcon parent must be the main window.

This is solved by making the system tray a child of the main
window. However, this approach has the following limitation: as the main
window is created once the UI has been built and this process uses the
QML engine, the 'systemTray' context property won't be available to
other QML components this stage. This context property will be set after
both the main UI and the system tray have been created.