Set a timeout on the process event connector socket
The process monitor thread just blocks until a new message with a
process event is received through the process event connector socket.

On systems that use kernels in which the option CONFIG_PROC_EVENTS has
not been set, no process events will be ever received through the
socket, thus leaving the process monitor in a perpetual blocked state
and preventing the helper from exiting when an exit command is received
from the application.

This problem is solved by setting a timeout on the socket. This way the
process monitor can react to the stop signal once the timeout has expired.

Fixes #71.