Save main window geometry when it changes
Saving the main window geometry when the application quits won't work
when the user logout from the session on window managers that just kill
the applications.

With this change, the window geometry will be saved after a short period
of time since the last received window geometry changed event.

See #312.