Fix CPU performance scaling mode not being restored from file
Call the right function to restore the mode from the saved file.

---
C++ rant: This is an example of the implicit type conversions footgun.

A copy/paste coupled with an human error shows how bad implicit type
conversions can be. The sane behaviour should have been a compilation
error. But no! Let's automatically convert it to a bool and call it a
day.
---

Fixes #322.