Fix compilation issues with Linux 6.0 api headers
In Linux 6.0, zero-length arrays were replaced with flexible-array
members.

Flexible-array members allows struct data types to end with an array
with no specified size. Although this is a C99 feature not yet supported
by the C++ standard, the compiler might enforce its proper use when used
in C++ code, as it's the case when the application is compiled using the
new Linux 6.0 API headers.

Fixes #325.