CONTRIBUTING: Require merge requests created from new branches
Merge requests created from the master branch do not allow maintainers
to contribute to the branch. This increases the noise level of the code
review and slows down the process of getting it merged on the master
branch.

Most people don't pay attention or are not aware of this detail, but
it's an important one, so make it a requirement.